Move mmap stuff out of core.py
