+++ tests/test_multiprocess.py
@@ -12,7 +12,6 @@ if sys.version_info < (2, 7):
 else:
     import unittest
 
-
 from prometheus_client import core
 from prometheus_client.core import (
     CollectorRegistry,
@@ -25,6 +24,7 @@ from prometheus_client.core import (
 from prometheus_client.multiprocess import (
     mark_process_dead,
     MultiProcessCollector,
+    _MultiProcessValue,
 )
 
 
@@ -32,7 +32,7 @@ class TestMultiProcess(unittest.TestCase):
     def setUp(self):
         self.tempdir = tempfile.mkdtemp()
         os.environ['prometheus_multiproc_dir'] = self.tempdir
-        core._ValueClass = core._MultiProcessValue(lambda: 123)
+        core._ValueClass = _MultiProcessValue(lambda: 123)
         self.registry = CollectorRegistry()
         self.collector = MultiProcessCollector(self.registry, self.tempdir)
 
@@ -43,7 +43,7 @@ class TestMultiProcess(unittest.TestCase):
 
     def test_counter_adds(self):
         c1 = Counter('c', 'help', registry=None)
-        core._ValueClass = core._MultiProcessValue(lambda: 456)
+        core._ValueClass = _MultiProcessValue(lambda: 456)
         c2 = Counter('c', 'help', registry=None)
         self.assertEqual(0, self.registry.get_sample_value('c_total'))
         c1.inc(1)
@@ -52,7 +52,7 @@ class TestMultiProcess(unittest.TestCase):
 
     def test_summary_adds(self):
         s1 = Summary('s', 'help', registry=None)
-        core._ValueClass = core._MultiProcessValue(lambda: 456)
+        core._ValueClass = _MultiProcessValue(lambda: 456)
         s2 = Summary('s', 'help', registry=None)
         self.assertEqual(0, self.registry.get_sample_value('s_count'))
         self.assertEqual(0, self.registry.get_sample_value('s_sum'))
@@ -63,7 +63,7 @@ class TestMultiProcess(unittest.TestCase):
 
     def test_histogram_adds(self):
         h1 = Histogram('h', 'help', registry=None)
-        core._ValueClass = core._MultiProcessValue(lambda: 456)
+        core._ValueClass = _MultiProcessValue(lambda: 456)
         h2 = Histogram('h', 'help', registry=None)
         self.assertEqual(0, self.registry.get_sample_value('h_count'))
         self.assertEqual(0, self.registry.get_sample_value('h_sum'))
@@ -76,7 +76,7 @@ class TestMultiProcess(unittest.TestCase):
 
     def test_gauge_all(self):
         g1 = Gauge('g', 'help', registry=None)
-        core._ValueClass = core._MultiProcessValue(lambda: 456)
+        core._ValueClass = _MultiProcessValue(lambda: 456)
         g2 = Gauge('g', 'help', registry=None)
         self.assertEqual(0, self.registry.get_sample_value('g', {'pid': '123'}))
         self.assertEqual(0, self.registry.get_sample_value('g', {'pid': '456'}))
@@ -88,7 +88,7 @@ class TestMultiProcess(unittest.TestCase):
 
     def test_gauge_liveall(self):
         g1 = Gauge('g', 'help', registry=None, multiprocess_mode='liveall')
-        core._ValueClass = core._MultiProcessValue(lambda: 456)
+        core._ValueClass = _MultiProcessValue(lambda: 456)
         g2 = Gauge('g', 'help', registry=None, multiprocess_mode='liveall')
         self.assertEqual(0, self.registry.get_sample_value('g', {'pid': '123'}))
         self.assertEqual(0, self.registry.get_sample_value('g', {'pid': '456'}))
@@ -102,7 +102,7 @@ class TestMultiProcess(unittest.TestCase):
 
     def test_gauge_min(self):
         g1 = Gauge('g', 'help', registry=None, multiprocess_mode='min')
-        core._ValueClass = core._MultiProcessValue(lambda: 456)
+        core._ValueClass = _MultiProcessValue(lambda: 456)
         g2 = Gauge('g', 'help', registry=None, multiprocess_mode='min')
         self.assertEqual(0, self.registry.get_sample_value('g'))
         g1.set(1)
@@ -111,7 +111,7 @@ class TestMultiProcess(unittest.TestCase):
 
     def test_gauge_max(self):
         g1 = Gauge('g', 'help', registry=None, multiprocess_mode='max')
-        core._ValueClass = core._MultiProcessValue(lambda: 456)
+        core._ValueClass = _MultiProcessValue(lambda: 456)
         g2 = Gauge('g', 'help', registry=None, multiprocess_mode='max')
         self.assertEqual(0, self.registry.get_sample_value('g'))
         g1.set(1)
@@ -120,7 +120,7 @@ class TestMultiProcess(unittest.TestCase):
 
     def test_gauge_livesum(self):
         g1 = Gauge('g', 'help', registry=None, multiprocess_mode='livesum')
-        core._ValueClass = core._MultiProcessValue(lambda: 456)
+        core._ValueClass = _MultiProcessValue(lambda: 456)
         g2 = Gauge('g', 'help', registry=None, multiprocess_mode='livesum')
         self.assertEqual(0, self.registry.get_sample_value('g'))
         g1.set(1)
@@ -130,13 +130,13 @@ class TestMultiProcess(unittest.TestCase):
         self.assertEqual(2, self.registry.get_sample_value('g'))
 
     def test_namespace_subsystem(self):
-         c1 = Counter('c', 'help', registry=None, namespace='ns', subsystem='ss')
-         c1.inc(1)
-         self.assertEqual(1, self.registry.get_sample_value('ns_ss_c_total'))
+        c1 = Counter('c', 'help', registry=None, namespace='ns', subsystem='ss')
+        c1.inc(1)
+        self.assertEqual(1, self.registry.get_sample_value('ns_ss_c_total'))
 
     def test_counter_across_forks(self):
         pid = 0
-        core._ValueClass = core._MultiProcessValue(lambda: pid)
+        core._ValueClass = _MultiProcessValue(lambda: pid)
         c1 = Counter('c', 'help', registry=None)
         self.assertEqual(0, self.registry.get_sample_value('c_total'))
         c1.inc(1)
@@ -149,7 +149,7 @@ class TestMultiProcess(unittest.TestCase):
     @unittest.skipIf(sys.version_info < (2, 7), "Test requires Python 2.7+.")
     def test_collect(self):
         pid = 0
-        core._ValueClass = core._MultiProcessValue(lambda: pid)
+        core._ValueClass = _MultiProcessValue(lambda: pid)
         labels = dict((i, i) for i in 'abcd')
 
         def add_label(key, value):
@@ -210,7 +210,7 @@ class TestMultiProcess(unittest.TestCase):
     @unittest.skipIf(sys.version_info < (2, 7), "Test requires Python 2.7+.")
     def test_merge_no_accumulate(self):
         pid = 0
-        core._ValueClass = core._MultiProcessValue(lambda: pid)
+        core._ValueClass = _MultiProcessValue(lambda: pid)
         labels = dict((i, i) for i in 'abcd')
 
         def add_label(key, value):
@@ -252,37 +252,6 @@ class TestMultiProcess(unittest.TestCase):
         ]
 
         self.assertEqual(metrics['h'].samples, expected_histogram)
-
-
-class TestMmapedDict(unittest.TestCase):
-    def setUp(self):
-        fd, self.tempfile = tempfile.mkstemp()
-        os.close(fd)
-        self.d = core._MmapedDict(self.tempfile)
-
-    def test_process_restart(self):
-        self.d.write_value('abc', 123.0)
-        self.d.close()
-        self.d = core._MmapedDict(self.tempfile)
-        self.assertEqual(123, self.d.read_value('abc'))
-        self.assertEqual([('abc', 123.0)], list(self.d.read_all_values()))
-
-    def test_expansion(self):
-        key = 'a' * core._INITIAL_MMAP_SIZE
-        self.d.write_value(key, 123.0)
-        self.assertEqual([(key, 123.0)], list(self.d.read_all_values()))
-
-    def test_multi_expansion(self):
-        key = 'a' * core._INITIAL_MMAP_SIZE * 4
-        self.d.write_value('abc', 42.0)
-        self.d.write_value(key, 123.0)
-        self.d.write_value('def', 17.0)
-        self.assertEqual(
-            [('abc', 42.0), (key, 123.0), ('def', 17.0)],
-            list(self.d.read_all_values()))
-
-    def tearDown(self):
-        os.unlink(self.tempfile)
 
 
 class TestUnsetEnv(unittest.TestCase):
